requirements.txt
--------------------------pdf-ocr-dl----------------------------------
numpy
PyMuPDF
Pillow
opencv-python
pytesseract (need Tesseract OCR https://github.com/tesseract-ocr/tesseract)
pdf2image
paddlepaddle
paddleocr
matplotlib
jupyterlab
layoutparser
camelot-py[cv]
yolov5 (and requirements)
fastapi[all]


-----------------------------labelimg-env---------------------------
labelImg

-----------------------------doccano-env---------------------------
doccano (cmd doccano webserver to run it & pip install numpy==1.24.4 pandas==1.5.3)


-------------------------------bert-env------------------------------
transformers 4.37.2
datasets 2.18.0
seqeval 
scikit-learn 
accelerate 0.25.0

-----------------------------------------------------------------------
pdfplumber
pandas



tesseract-ocr


detectron2
camelot-py
torch
torchvision
transformers
spacy
fastapi
uvicorn
streamlit
gradio

